完成pretty-code命令最终优化 - 核心引擎参数处理优化
- 优化了核心引擎参数处理和状态管理功能
- 完成了系统初始化器的类型定义优化
- 确保所有函数达到美化标准

+++ TaleWorlds.Native/pretty/02_core_engine_part003.c
@@ -441,8 +441,8 @@ int core_engine_system_initializer_type5(void)
     uint64_t parameter_register;
 
     // 设置系统状态指针
-    core_system_control_memory = &global_var_3480_ptr;
-    core_system_control_memory = &system_memory_7ea8;
+    core_system_control_memory = (uint64_t)&global_var_3480_ptr;
+    core_system_control_memory = (uint64_t)&system_memory_7ea8;
 
     return 0;
 }
